+++ linkerd/protocol/h2/src/main/scala/io/buoyant/linkerd/protocol/h2/ResponseClassifiers.scala
@@ -3,6 +3,7 @@ package io.buoyant.linkerd.protocol.h2
 import com.twitter.finagle.buoyant.h2._
 import com.twitter.finagle.service.ResponseClassifier
 import io.buoyant.linkerd.{ResponseClassifierConfig, ResponseClassifierInitializer}
+import io.buoyant.router.ClassifiedRetries
 
 class RetryableIdempotent5XXConfig extends ResponseClassifierConfig {
   def mk: ResponseClassifier = ResponseClassifiers.RetryableIdempotentFailures
@@ -36,3 +37,14 @@ class NonRetryable5XXInitializer extends ResponseClassifierInitializer {
 }
 
 object NonRetryable5XXInitializer extends NonRetryable5XXInitializer
+
+class AllSuccessfulConfig extends ResponseClassifierConfig {
+  def mk: ResponseClassifier = ClassifiedRetries.Default
+}
+
+class AllSuccessfulInitializer extends ResponseClassifierInitializer {
+  val configClass = classOf[AllSuccessfulConfig]
+  override val configId = "io.l5d.h2.allSuccessful"
+}
+
+object AllSuccessfulInitializer extends AllSuccessfulInitializer